backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
pydantic==2.12.5
python-multipart==0.0.22
google-generativeai
python-dotenv
